bit-endian fixes from Spencer Chang.

+++ elements/wifi/analysis/countfecbytes.hh
@@ -29,7 +29,7 @@ class CountFecBytes : public Element { public:
   int _packets;
 
   unsigned _length;
-  unsigned _et;
+  uint16_t _et;
   EtherAddress _src;
   int good_packet;
   bool _runs;
