+++ elements/wifi/frag/fragmentack.hh
@@ -52,7 +52,7 @@ class FragmentAck : public Element { public:
   FragTable _frags;
 
   EtherAddress _en;
-  unsigned _et;
+  uint16_t _et;
   unsigned _window_size;
 
   unsigned _ack_timeout_ms;
