+++ elements/wifi/sr/counterflood.hh
@@ -106,7 +106,7 @@ private:
 
   IPAddress _ip;    // My IP address.
   EtherAddress _en; // My ethernet address.
-  uint32_t _et;     // This protocol's ethertype
+  uint16_t _et;     // This protocol's ethertype
   IPAddress _bcast_ip;
 
   EtherAddress _bcast;
