+++ elements/wifi/sr/srqueryresponder.hh
@@ -51,7 +51,7 @@ class SRQueryResponder : public Element {
 
   IPAddress _ip;    // My IP address.
   EtherAddress _en; // My ethernet address.
-  uint32_t _et;     // This protocol's ethertype
+  uint16_t _et;     // This protocol's ethertype
